recent: Rename recent_topics_data.
