+++ web/tests/example5.test.js
@@ -25,7 +25,7 @@ const message_lists = mock_esm("../src/message_lists");
 const message_util = mock_esm("../src/message_util");
 const notifications = mock_esm("../src/notifications");
 const pm_list = mock_esm("../src/pm_list");
-const recent_topics_data = mock_esm("../src/recent_topics_data");
+const recent_view_data = mock_esm("../src/recent_view_data");
 const stream_list = mock_esm("../src/stream_list");
 const unread_ops = mock_esm("../src/unread_ops");
 const unread_ui = mock_esm("../src/unread_ui");
@@ -102,7 +102,7 @@ run_test("insert_message", ({override}) => {
     helper.redirect(message_util, "add_new_messages_data");
     helper.redirect(message_util, "add_new_messages");
     helper.redirect(notifications, "received_messages");
-    helper.redirect(recent_topics_data, "process_message");
+    helper.redirect(recent_view_data, "process_message");
     helper.redirect(stream_list, "update_streams_sidebar");
     helper.redirect(unread_ops, "process_visible");
     helper.redirect(unread_ui, "update_unread_counts");
@@ -124,7 +124,7 @@ run_test("insert_message", ({override}) => {
         [unread_ops, "process_visible"],
         [notifications, "received_messages"],
         [stream_list, "update_streams_sidebar"],
-        [recent_topics_data, "process_message"],
+        [recent_view_data, "process_message"],
     ]);
 
     // Despite all of our stubbing/mocking, the call to
