+++ web/tests/recent_topics.test.js
@@ -193,7 +193,7 @@ const {all_messages_data} = zrequire("all_messages_data");
 const people = zrequire("people");
 const rt = zrequire("recent_view_ui");
 const recent_view_util = zrequire("recent_view_util");
-const rt_data = zrequire("recent_topics_data");
+const rt_data = zrequire("recent_view_data");
 const muted_users = zrequire("muted_users");
 
 people.add_active_user({
